+++ bootique-agrest/src/test/java/io/bootique/agrest/AgrestModuleExtenderIT.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package io.bootique.linkrest;
+package io.bootique.agrest;
 
 import io.agrest.AgFeatureProvider;
 import io.agrest.AgModuleProvider;