+++ RELEASE-NOTES.md
@@ -4,6 +4,7 @@
 * #31 Upgrade JAX-RS to 2.1
 * #32 Upgrade to LinkRest 2.13
 * #34 Added integration with Agrest.
+* #35 Package conflict between bootique-agrest module and old bootique-linkrest module
 
 ## 0.25